% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullButton.R
\name{fullButton}
\alias{fullButton}
\alias{pageButton}
\alias{multiButton}
\title{Add a button}
\usage{
fullButton(..., outline = FALSE, clear = FALSE)

pageButton(..., outline = FALSE, clear = FALSE)

multiButton(..., outline = FALSE, clear = FALSE)
}
\arguments{
\item{...}{any other element}

\item{outline, clear}{set outline or clear button.}
}
\description{
Add a button
}
\examples{
fullButton("button", href = "https://john-coene.com")

}
