% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCard.R
\name{fullCards}
\alias{fullCards}
\alias{fullCard}
\title{Add a card}
\usage{
fullCards(...)

fullCard(..., header = NULL, footer = NULL)
}
\arguments{
\item{...}{your cards, see \code{\link{fullCard}}.}
}
\description{
Created a card
}
