% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullGrid.R
\name{fullRow}
\alias{fullRow}
\alias{fullColumn}
\alias{pageRow}
\alias{pageColumn}
\title{Add grid elements}
\usage{
fullRow(...)

fullColumn(...)

pageRow(...)

pageColumn(...)
}
\arguments{
\item{...}{any element.}
}
\description{
Add grid elements
}
