% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPlot.R
\name{fullSectionPlot}
\alias{fullSectionPlot}
\alias{fullSlidePlot}
\title{Use plot as background}
\usage{
fullSectionPlot(outputId, ..., menu = NULL, center = FALSE,
  background = "rgba(255,255,255,0.5)")

fullSlidePlot(outputId, ..., menu = NULL, center = FALSE,
  background = "rgba(255,255,255,0.5)")
}
\arguments{
\item{outputId}{id passed to \code{\link[shiny]{plotOutput}}.}

\item{...}{any children element.}

\item{menu}{menu link.}

\item{center}{whether to center text horizontally.}

\item{background}{sets background to overlaying shiny inputs.}
}
\description{
Use plot as background
}
