% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipage.R
\name{multiPage}
\alias{multiPage}
\alias{multiSection}
\title{Setup a multiscroll page}
\usage{
multiPage(..., opts = NULL, menu = NULL)

multiSection(..., menu = NULL, center = FALSE)
}
\arguments{
\item{...}{any element.}

\item{opts}{a list of options, see details.}

\item{menu}{named vector of menu links.}
}
\description{
Setup a multiscroll page
}
\examples{
if(interactive()){
library(shiny)

ui <- multiPage(
  multiLeft(
    multiSection(
      h1("Multiple")
    )
  ),
  multiRight(
    "Scroll"
  )
)

server <- function(input, output){}

ShinyApp(ui, server)
}

}
