% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTheme.R
\name{fullTheme}
\alias{fullTheme}
\title{Set theme}
\usage{
fullTheme(theme)
}
\arguments{
\item{theme}{theme, see details.}
}
\description{
Set theme
}
\details{
Valid \code{theme}:
\itemize{
  \item{\code{aqua}}
  \item{\code{black}}
  \item{\code{blue}}
  \item{\code{dark2}}
  \item{\code{fuchsia}}
  \item{\code{gray}}
  \item{\code{green}}
  \item{\code{lime}}
  \item{\code{maroon}}
  \item{\code{navy}}
  \item{\code{olive}}
  \item{\code{orange}}
  \item{\code{purple}}
  \item{\code{red}}
  \item{\code{silver}}
  \item{\code{teal}}
  \item{\code{yellow}}
}
}
