% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullContainer.R
\name{fullContainer}
\alias{fullContainer}
\alias{pageContainer}
\alias{multiContainer}
\title{Add containers}
\usage{
fullContainer(...)

pageContainer(...)

multiContainer(...)
}
\arguments{
\item{...}{any element.}
}
\description{
Add a container.
}
\details{
Pass within your grid, sections or slides.
}
\examples{
if(interactive()){
  library(shiny)

  ui <- fullPage(
    center = TRUE,
    fullSection(
      h1("Containers")
    ),
    fullSection(
      h1("Without container"),
      plotOutput("plot1")
    ),
    fullSection(
      fullContainer(
        h1("With container"),
        plotOutput("plot2"),
        pre(
          code(
            "fullContainer()"
          )
        )
      )
    )
  )

  server <- function(input, output){
    output$plot1 <- renderPlot({
      plot(mtcars$wt, mtcars$mpg)
    })

    output$plot2 <- renderPlot({
      plot(mtcars$wt, mtcars$mpg)
    })
  }

  shinyApp(ui, server)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
