% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullContainer.R
\name{fullContainer}
\alias{fullContainer}
\alias{pageContainer}
\title{Add a container}
\usage{
fullContainer(...)

pageContainer(...)
}
\arguments{
\item{...}{any element.}
}
\description{
Add a container
}
