% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPage.R
\name{fullSection}
\alias{fullSection}
\alias{fullSlide}
\alias{pageSection}
\title{Sections and Slides}
\usage{
fullSection(..., menu = NULL, center = FALSE)

fullSlide(..., menu = NULL, center = FALSE)

pageSection(..., menu = NULL, center = FALSE)
}
\arguments{
\item{...}{any element.}

\item{menu}{menu link.}

\item{center}{whether to center text horizontally.}
}
\description{
Add sections and slides.
}
\examples{
if(interactive()){

library(shiny)

ui <- fullPage(
  opts = list(
    sectionsColor = c("#f3f3f3", "#4BBFC3")
  ),
  fullSection(h1("First section"), center = TRUE),
  fullSection(
    center = TRUE,
    fullSlide(h2("Slide 1")),
    fullSlide(h2("Slide 2"))
  )
)

server <- function(input, output){}

shinyApp(ui, server)
}

}
