% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPage.R
\name{fullSection}
\alias{fullSection}
\alias{fullSlide}
\title{Sections and Slides}
\usage{
fullSection(..., menu = NULL, center = FALSE)

fullSlide(..., menu = NULL, center = FALSE)
}
\arguments{
\item{...}{any element.}

\item{menu}{menu link.}

\item{center}{whether to center text horizontally.}
}
\description{
Sections and Slides
}
\examples{
if(interactive()){

library(shiny)

ui <- fullPage(
  opts = list(sectionsColor = c('#4BBFC3', '#7BAABE')),
  fullSection(h3("First"), center = TRUE),
  fullSection(
    center = TRUE,
    fullSlide(h3("Slide 1")),
    fullSlide(h3("Slide 2"))
  )
)

server <- function(input, output){}

shinyApp(ui, server)
}

}
