% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPage.R
\name{fullPage}
\alias{fullPage}
\alias{fullSection}
\alias{fullSlide}
\title{Setup page}
\usage{
fullPage(...)

fullSection(...)

fullSlide(...)
}
\description{
Setup page
}
\examples{
library(shiny)

ui <- fullPage(
  fullSection(
    tags$h1("fullPage.js meets Shiny")
  ),
  fullSection(
    plotOutput("plot")
  ),
  fullSection(
    fullRow(
      fullColumn(
        shiny::tags$h1("Column 1")
      ),
      fullColumn(
        plotOutput("hist")
      ),
      fullColumn(),
      fullColumn()
    )
  )
)

server <- function(input, output){
  output$plot <- renderPlot({
    hist(rnorm(20, 1, 5))
  })

  output$hist <- renderPlot({
    hist(rnorm(20, 1, 5))
  })
}

shinyApp(ui, server)

}
