% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullButton.R
\name{fullButton}
\alias{fullButton}
\title{Add a button}
\usage{
fullButton(..., outline = FALSE, clear = FALSE, tag = shiny::a())
}
\arguments{
\item{...}{any other element}

\item{outline, clear}{set outline or clear button.}

\item{tag}{button tag, generally \code{a} or \code{button}.}
}
\description{
Add a button
}
