% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullGrid.R
\name{fullRow}
\alias{fullRow}
\alias{fullColumn}
\alias{pageRow}
\alias{pageColumn}
\alias{multiRow}
\alias{multiColumn}
\title{Add grid elements}
\usage{
fullRow(...)

fullColumn(..., width = 6)

pageRow(...)

pageColumn(..., width = 6)

multiRow(...)

multiColumn(..., width = 6)
}
\arguments{
\item{...}{any element.}

\item{width}{Width of column.}
}
\description{
Add rows and columns.
}
\details{
Wrap columns in row function, see examples.
}
\examples{
if(interactive()){
  library(shiny)

  ui <- pagePiling(
    center = TRUE,
    sections.color = c("#f3f3f3", "#f4f4f4"),
    pageSection(
      h1("Grid")
    ),
    pageSection(
      h1("Columns"),
      pageRow(
        pageColumn(
          h2("First Column"),
          p("Resize your screen to see columns in action.")
        ),
        pageColumn(
          h2("Second Column"),
          numericInput("obs", "observations", value = 20)
        ),
        pageColumn(
          h2("Thirs Column"),
          plotOutput("plot")
        )
      )
    )
  )

  server <- function(input, output){
    output$plot <- renderPlot({
      par(bg = "#f4f4f4")
      hist(rnorm(input$obs, 10, 3))
    })
  }

  shinyApp(ui, server)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
