% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPage.R
\name{multiSection}
\alias{multiSection}
\alias{multiSectionImage}
\title{Create multi sections}
\usage{
multiSection(..., menu = NULL, align = "none")

multiSectionImage(..., img, side, menu = NULL, align = "none")
}
\arguments{
\item{...}{any element.}

\item{menu}{named vector of menu links.}

\item{align}{text alignment, takes \code{left}, \code{right} or, \code{center}, defaults to \code{none}.}

\item{img}{path to image.}

\item{side}{image side.}
}
\description{
Create multi sections.
}
\examples{
if(interactive()){
  library(shiny)

  options <- list(
    sectionsColor = c("#f3f3f3", "#d3d3d3", "#000")
  )

  ui <- multiPage(
    opts = options,
    multiLeft(
      align = "right",
      multiSection(
        h1("multi")
      ),
      multiSection(
        h2("Section ")
      ),
      multiSection(
        h2("Image ")
      )
    ),
    multiRight(
      align = "left",
      multiSection(
        h1("page()")
      ),
      multiSection(
        h2(" 2")
      ),
      multiSectionImage(
        img = "https://alvarotrigo.com/multiScroll/imgs/tiger.jpg",
        side = "right",
        h2("Background")
      )
    )
  )

  server <- function(input, output){}

  shinyApp(ui, server)
}

}
