% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagePiling.R
\name{pagePiling}
\alias{pagePiling}
\title{Set page piling}
\usage{
pagePiling(..., opts = NULL, menu = NULL)
}
\description{
Set page piling
}
\examples{
if(interactive()){
ui <- pagePiling(
  pageSection(
    h1("Page piling")
  ),
  pageSection(
    h1("Section2")
  )
)

}

}
