% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagePiling.R
\name{pagePiling}
\alias{pagePiling}
\title{Setup a page.}
\usage{
pagePiling(..., sections.color, opts = list(), menu = NULL,
  center = FALSE)
}
\arguments{
\item{...}{any element.}

\item{sections.color}{character vector, i.e.: \code{c("blue", "red")}}

\item{opts}{list of options, see details.}

\item{menu}{menu links as named vector.}

\item{center}{horizontally center all sections.}
}
\description{
Create a full scrolling page.
}
\details{
\itemize{
  \item{\code{direction} - takes \code{vertical} or \code{horizontal} - defines if the scroll is vertical or horizontal.}
  \item{\code{verticalCentered} - boolean.}
  \item{\code{scrollingSpeed} - scrolling transitions in milliseconds, numeric value i.e.: \code{700}.}
  \item{\code{easing} - character string - takes \code{swing} or \code{linear}.}
  \item{\code{loopBottom} - boolean - defines whether scrolling down in the last section should scroll to the first one or not.}
  \item{\code{loopTop} - boolean - defines whether scrolling up in the first section should scroll to the last one or not.}
  \item{\code{css3} - boolean - defines whether to use JavaScript or CSS3 transforms to scroll within sections. Useful to speed up the movement in tablet and mobile devices with browsers supporting CSS3. If this option is set to true and the browser doesn't support CSS3, a jQuery fallback will be used instead.}
  \item{\code{navigation} - boolean - if set, it will show a navigation bar made up of small circles.}
  \item{\code{touchSensitivity} - numeric value - defines a percentage of the browsers window width/height, and how far a swipe must measure for navigating to the next section.}
  \item{\code{keyboardScrolling} - boolean - defines if the content can be navigated using the keyboard.}
  \item{\code{animateAnchor} - boolean - defines whether the load of the site when given an anchor will scroll with animation to its destination or will directly load on the given section.}
}
}
\examples{
if(interactive()){
# Run de demo
demo("pagePiling", package = "fullPage")

# Or
library(shiny)

options <- list(
  loopBottom = TRUE
)

ui <- pagePiling(
  center = TRUE,
  sections.color = c('#f2f2f2', '#2C3E50', '#39C'),
  opts = options,
  menu = c("Section 1" = "section1",
           "Piling" = "section2",
           "Plots" = "section3",
           "Reactive" = "section4",
           "Images" = "images"),
  pageSection(
    menu = "section1",
    h1("Page piling")
  ),
  pageSection(
    menu = "section2",
    h1("Section 2")
  ),
  pageSectionPlot(
    "plot",
    menu = "section3",
    h1("Plot background")
  ),
  pageSectionPlot(
    "plot2",
    menu = "section4",
    pageContainer(
      h1("Reactive background"),
      sliderInput(
        "bins",
        "Data Points",
        min = 100,
        max = 500,
        step = 25,
        value = 200
      )
    )
  ),
  pageSectionImage( # will not show in viewer, open in browser
    menu = "image",
    img = paste0(
      "https://raw.githubusercontent.com/alvarotrigo/",
      "fullPage.js/master/examples/imgs/bg2.jpg"),
    h1("Image background")
  )
)

server <- function(input, output){

  output$plot <- renderPlot({
    par(bg = "grey60")
    plot(mtcars$wt, mtcars$mpg)
  })

  output$plot2 <- renderPlot({
    par(bg = "grey80")
    hist(rnorm(input$bins, 100, 250))
  })
}

shinyApp(ui, server)
}


}
\seealso{
\href{Official documentation}{https://github.com/alvarotrigo/pagePiling.js/}.
}
