% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSides.R
\name{multiLeft}
\alias{multiLeft}
\alias{multiRight}
\title{Create multi sections}
\usage{
multiLeft(..., align = NULL)

multiRight(..., align = NULL)
}
\arguments{
\item{...}{any element.}

\item{align}{text alignement passed to \code{CSS}, takes \code{left}, \code{right} or \code{center}.}
}
\description{
Create multi page sides.
}
\details{
Only use one of each in \code{\link{multiPage}}.
}
