% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSides.R
\name{multiLeft}
\alias{multiLeft}
\alias{multiRight}
\title{Create multi sections}
\usage{
multiLeft(...)

multiRight(...)
}
\arguments{
\item{...}{any element.}
}
\description{
Create multi page side.
}
\details{
Only use one of each in \code{\link{multiPage}}.
}
