% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTheme.R
\name{fullTheme}
\alias{fullTheme}
\alias{multiTheme}
\alias{pageTheme}
\title{Set theme}
\usage{
fullTheme(theme)

multiTheme(theme)

pageTheme(theme)
}
\arguments{
\item{theme}{theme, see details.}
}
\description{
Set a theme.
}
\details{
Valid \code{theme}:
\itemize{
  \item{\code{aqua}}
  \item{\code{black}}
  \item{\code{blue}}
  \item{\code{dark2}}
  \item{\code{fuchsia}}
  \item{\code{gray}}
  \item{\code{green}}
  \item{\code{lime}}
  \item{\code{maroon}}
  \item{\code{navy}}
  \item{\code{olive}}
  \item{\code{orange}}
  \item{\code{purple}}
  \item{\code{red}}
  \item{\code{silver}}
  \item{\code{teal}}
  \item{\code{yellow}}
}
}
\examples{
if(interactive()){

  ui <- fullPage(

    center = TRUE,

    fullTheme("red"), # theme

    fullSection(
      center = TRUE,
      h1("Themes")
    ),
    fullSection(
      fullButtonTo("Move to first section", section = 1)
    ),
    fullSection(
      center = TRUE,
      fullSlide(
        fullButtonRight("Slide right", clear = TRUE)
      ),
      fullSlide(
        fullButtonLeft("Slide left", clear = TRUE)
      )
    )
  )

  server <- function(input, output){}

  shinyApp(ui, server)
}
}
\author{
John Coene, \email{jcoenep@gmail.com}
}
